mem: finish the measured-accounting seam -- /sys/mem/used width + free command
Two readers were left behind by the measured-accounting change (a251d40),
which widened tiku_process_sram_used()/fram_used() to uint32_t (declared
proc-mem plus the live arena, e.g. a BASIC program's hundreds of KB):

  * /sys/mem/used summed those uint32_t values into a uint16_t accumulator,
    so it reported the total modulo 65536 -- the very number a251d40 set out
    to un-lie still wrapped. Widen the accumulator to uint32_t.
  * the `free` command's process table still printed the raw 16-bit
    p->sram_used/fram_used with %4u, so it showed a BASIC-heavy process as
    ~0 while ps and /proc (already on the helpers) showed the real footprint.
    Route it through tiku_process_sram_used()/fram_used() with %6lu, matching
    ps, and widen its columns.

HW-verified on the apollo510b Blue with a live ~273 KB BASIC arena: all three
readers now agree -- /sys/mem/used = 273000 (the old uint16 path would have
shown 10856), ps and free both report 273000. Generic kernel files; builds
clean on apollo510b, apollo4l and rp2350.

+++ kernel/vfs/tree/tiku_vfs_tree_sys.c
@@ -362,12 +362,15 @@ mem_free_read(char *buf, size_t max)
 /**
  * @brief Read handler for /sys/mem/used.
  *
- * Renders the sum of every registered process's sram_used
- * accounting field as a decimal line.  This measures what
- * processes *declared* through the proc-mem API, not actual SRAM
- * consumption — kernel statics and the stack are not included.
- * Walks the registry by pid; empty slots return NULL from
- * tiku_process_get() and are skipped.
+ * Renders the sum of every registered process's SRAM use as a decimal
+ * line: each process's declared proc-mem footprint plus its measured
+ * live allocation (e.g. the BASIC arena), via tiku_process_sram_used().
+ * Kernel statics and the stack are not included.  Walks the registry by
+ * pid; empty slots return NULL from tiku_process_get() and are skipped.
+ *
+ * The accumulator is 32-bit: tiku_process_sram_used() returns a uint32_t
+ * (a BASIC arena alone is hundreds of KB on the big-RAM parts), so a
+ * 16-bit sum here would wrap modulo 65536 and under-report.
  *
  * @param buf  Output buffer for the rendered text
  * @param max  Capacity of @p buf in bytes
@@ -376,7 +379,7 @@ mem_free_read(char *buf, size_t max)
 static int
 mem_used_read(char *buf, size_t max)
 {
-    uint16_t total = 0;
+    uint32_t total = 0;
     uint8_t i;
     for (i = 0; i < TIKU_PROCESS_MAX; i++) {
         struct tiku_process *p = tiku_process_get((int8_t)i);
